añadir parámetro w + verbose data

+++ problems/clp/evaluators/VCS_Function.h
@@ -52,7 +52,7 @@ public:
 	}
 
 
-	long cs[6];
+	double cs[6];
 
 	static int nn;
